styles,pages and most of the conections done

+++ src/app/features/admin/components/tables/colective-table/colective-table.component.ts
@@ -1,4 +1,6 @@
+import { DeleteAlertComponent } from '../../alerts/delete-alert/delete-alert.component';
 import { Component, Input, OnInit, OnChanges, SimpleChanges, ChangeDetectorRef, ChangeDetectionStrategy } from '@angular/core';
+import { ConfirmDeleteAlertComponent } from '../../alerts/confirm-delete-alert/confirm-delete-alert.component';
 import { CommonModule } from '@angular/common';
 import { ColectiveRepository } from '../../../data/repository/colective-repository';
 import { Colective } from '../../../data/models/colective.model';
@@ -6,12 +8,15 @@ import { Colective } from '../../../data/models/colective.model';
 @Component({
   selector: 'app-colective-table',
   standalone: true,
-  imports: [CommonModule],
+  imports: [CommonModule, ConfirmDeleteAlertComponent, DeleteAlertComponent],
   templateUrl: './colective-table.component.html',
   styleUrls: ['./colective-table.component.scss'],
   changeDetection: ChangeDetectionStrategy.Default
 })
 export class ColectiveTableComponent implements OnInit, OnChanges {
+  showConfirmDelete = false;
+  showDeleteAlert = false;
+  indexToDelete: number | null = null;
   ngOnChanges(changes: SimpleChanges): void {
     if (changes['token'] && !changes['token'].firstChange) {
       this.loadColectivos();
@@ -52,18 +57,36 @@ export class ColectiveTableComponent implements OnInit, OnChanges {
     });
   }
 
-  eliminar(index: number) {
-    const id = this.colectivos[index].id;
-    const tokenToUse = this.getToken();
-    console.log('[ColectiveTable] delete called with id:', id, 'and token:', tokenToUse);
-    this.repository.delete(id, tokenToUse).subscribe({
-      next: (res) => {
-        console.log('[ColectiveTable] delete success:', res);
-        this.loadColectivos();
-      },
-      error: (err) => {
-        console.error('[ColectiveTable] Error al eliminar colectivo', err);
-      }
-    });
+  onDeleteClick(index: number) {
+    this.indexToDelete = index;
+    this.showConfirmDelete = true;
+  }
+
+  onCancelDelete() {
+    this.showConfirmDelete = false;
+    this.indexToDelete = null;
+  }
+
+  onConfirmDelete() {
+    if (this.indexToDelete !== null) {
+      const id = this.colectivos[this.indexToDelete].id;
+      const tokenToUse = this.getToken();
+      this.repository.delete(id, tokenToUse).subscribe({
+        next: (res) => {
+          this.loadColectivos();
+          this.showConfirmDelete = false;
+          this.indexToDelete = null;
+          this.showDeleteAlert = true;
+        },
+        error: (err) => {
+          this.showConfirmDelete = false;
+          this.indexToDelete = null;
+        }
+      });
+    }
+  }
+
+  onCloseDeleteAlert() {
+    this.showDeleteAlert = false;
   }
 }
